HOCS-2457 add spring profile switching to swagger
