+++ src/main/java/uk/gov/digital/ho/hocs/casework/application/SwaggerConfiguration.java
@@ -1,0 +1,32 @@
+package uk.gov.digital.ho.hocs.casework.application;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@Slf4j
+@EnableSwagger2
+@Profile({"swagger"})
+public class SwaggerConfiguration implements WebMvcConfigurer {
+
+    @Bean
+    public Docket api() {
+        log.info("Swagger enabled in Casework Application...");
+        return new Docket(DocumentationType.SWAGGER_2)
+                .apiInfo(new ApiInfo("DECS Casework API","Casework Endpoints", "", "", "","",""))
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.any())
+                .build();
+    }
+
+}
